initial sinatra commit
