+++ Gemfile
@@ -1,0 +1,4 @@
+source 'https://rubygems.org'
+
+gem 'sinatra'
+gem 'sinatra-contrib'
