+++ lib/player.rb
@@ -1,0 +1,6 @@
+class Player 
+    def initialize
+        puts "What is your name?"
+        @name = gets.chomp.upcase
+    end
+end
