+++ main.rb
@@ -1,0 +1,5 @@
+require './lib/game'
+require './lib/mastermind'
+require './lib/player'
+
+Mastermind.new
